Fix Expand / Contract of the "Show" button (#237)
* Fix Expand / Contract of the "Show" button

Now, when switching between List and Tabular view,
the "Show" button will not expand, contract, move,
or show any animation. Basically the button animation
was disabled for this scenario.

Also, making the height of the button consistent
among the two different views

Fixes XP-2294

* Build package

* Remove units from `0`s

* Build app

* Revert `s` unit removal from `transition`

The `s` or `ms` unit is necessary for the CSS property
to work. If removed, it'll become an invalid property

* Disable transition for buttons

* Build app