"List Refactor" version release (#241)
* Add ability to not render the status circle if no status is passed (#236)

* Add avility to not render the status circle if no status is passed

* Add a new use case for the component

* Update snapshots

* Fix on clear method on select box (#239)

* Fix on clear method on select box

* Code refactor

* Build app

* Fix AdminScreen z-index (#238)

* adjust z-index for AdminScreen header (#2291)

* Update snapshots

* Fix Expand / Contract of the "Show" button (#237)

* Fix Expand / Contract of the "Show" button

Now, when switching between List and Tabular view,
the "Show" button will not expand, contract, move,
or show any animation. Basically the button animation
was disabled for this scenario.

Also, making the height of the button consistent
among the two different views

Fixes XP-2294

* Build package

* Remove units from `0`s

* Build app

* Revert `s` unit removal from `transition`

The `s` or `ms` unit is necessary for the CSS property
to work. If removed, it'll become an invalid property

* Disable transition for buttons

* Build app

* Stretch divider line below "Select List" input (#234)

* Stretch divider line below "Select List" input

This stretches the divider line between the "Select List" input
and the rest of the inputs in the SearchForm component.

Closes XP-2284

* Update `SearchForm` component snapshots

* npm run build

* Remove empty declaration

* Fix Keywords select's suggestions width and position

* Update snapshots