Fix `closeDropdown` prop in `SelectBox`
The `closeDropdown` prop from the `SelectBox` component was designed
so that it would only work when it was provided by a parent
`Dropdown` component:

```js
<Dropdown>
  <SelectBox closeDropdown />
</Dropdown>
```

However, when using it outside a `Dropdown` this has no effect on
the `SelectBox`.

This commit fixes the behavior by using the `closeDropdown` prop
from the internal `Dropdown` component inside `SelectBox` when
it isn't provided by any parent.