Make the Header bar appear above all elements on profile pages (#506)
* Rebuild the app

* Rebuild the app

* Make the header bar appear above all elements on the profile page

Set the z-index to the HeaderBar component to 1000. This is greater than
every UI component that appears on that page.

Remove the z-index from the element that contains the HeaderBar on the
ApplicantInfo page. Setting a z-index on that element creates a new
"stacking context" in CSS which makes the z-index for contained
elements do the wrong thing.

* Move z-index to the container rather than the HeaderBar

Co-authored-by: Nick Plekhanov <nickatgit@yandex.com>
# Conflicts:
#	lib/all.js
#	lib/all.js.map
#	lib/registration.js
#	lib/registration.js.map