Admin work experience styles (#545)
* modified styles for better integration with xp admin page

* Add styles to the WorkExperienceEntryPresenter stories

Since the WorkExperienceEntryPresenter no longer has its own text
styles, we need to create a wrapper to avoid giving the impression that
its text styling is broken.

* Hide work experience items beyond the 3rd and make buttons wider

In accordance with the spec, only 3 work experience entries are added at
the start, with an option to view more.

Buttons are changed into a `wide outline` style to match the image.

* fix mistaken use of class => className

* Update src/components/ui/WorkExperience/index.js

Formatting jsx a little more nicely.

Co-authored-by: Bernardo Dias <bernardodiasdacruz@gmail.com>

* fix things introduced by making changes in the github UI

* Fix CustomSelector styles

* minor improvements to formatting and api

Co-authored-by: Bernardo Dias <bernardodiasdacruz@gmail.com>